Update SettingsController.cs
Add Method to generate the Settings_Names on the first load. When the module view is loaded for the first time, the Settings_Names are checked in the database and if they do not exist, an error is generated in the view. I have added this method to correct this error. The first time the page is loaded these are automatically created with the value false.