Update PagesController.cs
I have changed for this Controller the method that I had previously added in SettingsController.cs. It is only executed the first time the module is loaded the first time to generate the Settings_Names. I have also modified one of the conditions so that it is only executed in the case that these Settings_Names already exist in the database.