Update package.json
I noticed that in the package.json file the production environment is not referenced and I have added it. This for when we do the angular build, the generated files are the files in release mode and do not have the numbers that appear before the point as in development.