Update PagesController.cs
I modified the GetPagesList method to get the total number of pages on the site by storing it in the allPages variable. I added a new parameter to it (getAllPages:true) to reuse the existing method to get the paginated result.
I removed the implemented method to filter pages missing metadata to reuse the same implementation to get the paginated result and total pages. The new method is created in PagesControllerImpl called GetPagesMissingMetadata.